requirements.txt
opencv-python==4.6.0.66
boto3==1.24.75
Flask==2.2.0
requests==2.25.1
xmltodict==0.12.0
mysql-connector-python==8.0.31
